Update README.md
It's still wrong as hell but this is the first commit.  These are the formulas pasted raw from Swift into the Haskell functions.  

I still have to translate the syntax into proper Haskell.  At the moment, it compiles but I'm not sure how correct it actually is.  

Loving Haskell so far.  Amazingly mind-opening!